Use derives for Input type
