+++ rumbas_support_derive/src/overwrite.rs
@@ -27,8 +27,7 @@ pub struct OverwriteReceiver {
     /// Receives the body of the struct or enum. We don't care about
     /// struct fields because we previously told darling we only accept structs.
     data: ast::Data<InputVariantReceiver, InputFieldReceiver>,
-    attrs: Vec<syn::Attribute>,
-
+    //attrs: Vec<syn::Attribute>,
     /// I guess we can't get other derives into `attrs` so we have to create our
     /// own derive list.
     //derive: darling::util::PathList,
@@ -42,38 +41,12 @@ impl ToTokens for OverwriteReceiver {
             ref ident,
             ref generics,
             ref data,
-            ref attrs,
+            //  ref attrs,
             ref input_name,
             //ref derive,
         } = *self;
 
-        println!("{:?}", ident);
-        let derive_attrs = attrs
-            .iter()
-            .filter(|a| a.path.is_ident("derive"))
-            .collect::<Vec<_>>();
-        println!("{:#?}", derive_attrs);
-        let to_derive = derive_attrs
-            .iter()
-            .flat_map(|a| {
-                let group = a.tokens.clone().into_iter().next().expect("Invalid derive");
-                match group {
-                    proc_macro2::TokenTree::Group(g) => g
-                        .stream()
-                        .into_iter()
-                        .flat_map(|tt| {
-                            if let proc_macro2::TokenTree::Ident(i) = tt {
-                                vec![i]
-                            } else {
-                                vec![]
-                            }
-                        })
-                        .collect::<Vec<_>>(),
-                    _ => panic!("Invalid derive"),
-                }
-            })
-            .collect::<Vec<_>>();
-        println!("{:?}", to_derive);
+        //println!("{:?}", ident);
 
         let input_ident = syn::Ident::new(&input_name, ident.span());
 
@@ -258,24 +231,4 @@ fn overwrite_handle_enum_variants(
             }
         })
         .collect::<Vec<_>>()
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    #[test]
-    fn used_to_debug() {
-        let derive_input = syn::parse_str(
-            r#"/// Some comment
-            #[derive(Clone, Input, Overwrite)]
-            #[input(name="test")]
-            pub struct A {
-                /// Hi
-                d: bool
-            }"#,
-        )
-        .unwrap();
-        let parsed = OverwriteReceiver::from_derive_input(&derive_input).unwrap();
-        assert_eq!(false, true);
-    }
 }
