+++ rumbas_support_derive/tests/mixed_enum.rs
@@ -10,8 +10,9 @@ use rumbas_support::overwrite::Overwrite;
 use rumbas_support::value::Value;
 use serde::Deserialize;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "TestInput")]
+#[derive(Clone, Deserialize)]
 pub enum Test {
     Unit,
     Tuple(f64, bool, String),
@@ -20,8 +21,9 @@ pub enum Test {
 
 type TestInputs = Vec<Test>;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "Test2Input")]
+#[derive(Clone, Deserialize)]
 pub struct Test2 {
     field1: TestInputs,
     field2: f64,
