+++ rumbas_support_derive/tests/tuple_struct.rs
@@ -8,18 +8,21 @@ use rumbas_support::overwrite::Overwrite;
 use rumbas_support::value::Value;
 use serde::Deserialize;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "TestInput")]
+#[derive(Clone, Deserialize)]
 pub struct Test(bool, f64);
 
 type TestInputs = Vec<Test>;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "Test2Input")]
+#[derive(Clone, Deserialize)]
 pub struct Test2(TestInputs, f64);
 
-#[derive(Clone, Input, Deserialize, Overwrite)]
+#[derive(Input, Overwrite)]
 #[input(name = "TestOverwriteInput")]
+#[derive(Clone, Deserialize)]
 pub struct TestOverwrite(bool, f64);
 
 #[cfg(test)]
