+++ rumbas_support_derive/tests/unit_enum.rs
@@ -8,8 +8,9 @@ use rumbas_support::overwrite::Overwrite;
 use rumbas_support::value::Value;
 use serde::Deserialize;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "TestInput")]
+#[derive(Clone, Deserialize)]
 pub enum Test {
     First,
     Second,
@@ -17,20 +18,21 @@ pub enum Test {
 
 type TestInputs = Vec<Test>;
 
-#[derive(Clone, Deserialize, Input)]
+#[derive(Input)]
 #[input(name = "Test2Input")]
+#[derive(Clone, Deserialize)]
 pub struct Test2 {
     field1: TestInputs,
     field2: f64,
 }
 
-/*
-#[derive(Clone, Input, Deserialize, Overwrite)]
+#[derive(Input, Overwrite)]
 #[input(name = "TestOverwriteInput")]
+#[derive(Clone, Deserialize)]
 pub struct TestOverwrite {
     field1: bool,
     field2: f64,
-}*/
+}
 
 #[cfg(test)]
 mod test {
